requirements.txt
Django==5.0.4
djangorestframework==3.15.1
psycopg2-binary==2.9.9
python-decouple==3.8
stripe==9.6.0
boto3==1.34.0
awsgi==0.2.7